Complete code coverage for PrimaryCleanPipeline
Code coverage is as close to 100% as possible. There is an
under-the-hood oddity that will not let me achieve 100% code coverage
due to the statement:
'else if (meanConfidence <= 70.00)'

For more information on this one can see the example in this online
forum: https://social.msdn.microsoft.com/Forums/en-US/5f7eb318-9ca4-4914-b6a5-7447f4b98ce8/partially-touched-code-question

Also added an ArgumentException throw to my test case to cover the
try-catch block in Execute function.