Add test read of an unclassified ingredient
This test confirms that our system will store new ingredients (ie.
ingredients that do not exist in any of our lists:
IngredientDictionary.txt, NonVeganIngredients.txt, or
UnclassifiedIngredients.txt). Once a new ingredient encountered, it will
be added to UnclassifiedIngredients.txt to be so that it may be
classified at a later time by an authorized user.