Check ingredient classification local file (#12)
* Create IngredientLocalRepository.cs

Will perform ingredient data access when using local files in place of a database.

 Will ultimately convert to using a database.

* Add more test photos

* Update testing architecture

* Add ExtractDualNamedIngredients

--Fixed bug in function (now works across entire test photo set)
-- Completed test collection

* Update NonVeganIngredients.txt

Converted all ingredients to lower-case

* Add test refrence to web api media folder

Now uses dependency injection to perform a read for an ingredient to
determine that it is vegan.

* Add collection to testing project

Created IngredientLocalRepositoryCollection.cs to aid in the unit
testing of IngredientLocalRepository.cs (in the Is This Vegan (Net)
project).

This file's contents are not currently complete; it will be completed as
tests are being written.

* Add single non-vegan ingredient read test

Tests a read for a singular non-vegan ingredient

* Add [Get/Reset]InitialList helper methods

GetInitialList reads an ingredient dictionary from web api project into memory. ResetInitialList will reset the ingredient dictionary to its contents before having run the test.

* Add skim milk to NonVeganIngredients.txt

Ingredient was not categorized correctly

* Add MaybeVegan ingredient read test

* Update IngredientClassificationEnum

Changed 'NotRecognized' to 'Unclassified' to avoid confusion.

* Add test read of an unclassified ingredient

This test confirms that our system will store new ingredients (ie.
ingredients that do not exist in any of our lists:
IngredientDictionary.txt, NonVeganIngredients.txt, or
UnclassifiedIngredients.txt). Once a new ingredient encountered, it will
be added to UnclassifiedIngredients.txt to be so that it may be
classified at a later time by an authorized user.

* Refactor GetList test helper method

Uses parameter instead of hard-coded filename value.

* Implement Create method

Moved implementation of AddToUnclassified into Create because they are redundant.

* Add IsUnclassified

* Add Update unit test

Tests for NotSupportedException (this will not be available in the first
version of the app)

* Add Delete unit test

Tests for NotSupportedException (this will not be available in the first
version of the app)