Fix '.Contains()' bug
Was getting wrong values when checking ingredient dictionaries due to
the fact that (string)variable.Contains(target) will return true if the
target is present anywhere in the string. For example, if we search for
a target, "sugar" in NonVeganIngredients.txt, contains will return true
due to the fact that "lactose - milk sugar" exists in it.

To fix this, a string.Split('\n') operation is performed on the read
file prior to the .Contains(target) call.