Secondary api (#13)
* Update C# to 8.0

* Add folder and class for CategorizeBackend tests

Creates 'Categorize' folder in Backend
Creates 'CategorizeBackendTest.cs' file in Categorize folder

* Add UpdateCategory method

Updates the category of a product as a result of the category of a given
ingredient.

* Fix '.Contains()' bug

Was getting wrong values when checking ingredient dictionaries due to
the fact that (string)variable.Contains(target) will return true if the
target is present anywhere in the string. For example, if we search for
a target, "sugar" in NonVeganIngredients.txt, contains will return true
due to the fact that "lactose - milk sugar" exists in it.

To fix this, a string.Split('\n') operation is performed on the read
file prior to the .Contains(target) call.

* Add CheckIngredients unit test

Correctly retrieves the categories of ingredients from local files, then
correctly categorizes a product based on the categories of the
ingredients.

* Add CategorizeProduct and unit tests

Takes a user-authenticated string of a product's ingredient list, then
checks each ingredient against the local ingredient dicitonary, and
finally categorizes the product based on the categories of its
ingredients.

* Create Post api for user-authenticated list

Provides the endpoint that receives a user-authenticated ingredient list
for a product. Once validated, this list is given to CategorizeBackend
which will check each ingredient to the local ingredient dictionary,
thereby enabling the system to categorize the product. This result is
then returned to the end-user.

* Fix bugs after test with Postman

- Gave exceptions variable names to aid in debugging
- Fixed bug resulting from string.Split(). When running on Azure server,
  newlines are denoted by \r\n, and not just \n.

* Add new file 'profile.arm.json'